[+] refactore name global function
