+++ includes/global/global.hpp
@@ -1,9 +1,8 @@
 #ifndef JS_CMP_GLOBAL_HPP
 #define JS_CMP_GLOBAL_HPP
 
-#include "global_isFinite.hpp"
+#include "isFinite.hpp"
 #include "global_isNaN.hpp"
-#include "global_parseFloat.hpp"
 #include "global_parseInt.hpp"
 #include "global_typeof.hpp"
 #include "global_void.hpp"
