--- includes/global/global_isFinite.hpp
@@ -1,10 +1,0 @@
-#ifndef JS_CMP_GLOBAL_ISFINITE_HPP
-#define JS_CMP_GLOBAL_ISFINITE_HPP
-
-#include "../types/JsAny.hpp"
-#include "../utils/Convert.hpp"
-
-#include <limits>
-bool global_isFinite(const JS::Any& number);
-
-#endif // JS_CMP_GLOBAL_ISFINITE_HPP
