--- includes/global/global_isNaN.hpp
@@ -1,10 +1,0 @@
-#ifndef JS_CMP_GLOBAL_ISNAN_HPP
-#define JS_CMP_GLOBAL_ISNAN_HPP
-
-#include "../types/JsAny.hpp"
-#include "../utils/Convert.hpp"
-
-#include <limits>
-bool global_isNaN(const JS::Any& number);
-
-#endif // JS_CMP_GLOBAL_ISNAN_HPP
