--- includes/global/global_typeof.hpp
@@ -1,7 +1,0 @@
-#ifndef JS_CMP_GLOBAL_TYPEOF_HPP
-#define JS_CMP_GLOBAL_TYPEOF_HPP
-
-#include "../includes/types/JsAny.hpp"
-std::string global_typeof(const JS::Any any);
-
-#endif // JS_CMP_GLOBAL_TYPEOF_HPP
