--- includes/global/global_void.hpp
@@ -1,7 +1,0 @@
-#ifndef JS_CMP_GLOBAL_VOID_HPP
-#define JS_CMP_GLOBAL_VOID_HPP
-
-#include "../includes/types/JsAny.hpp"
-JS::Undefined global_void(const JS::Any& args);
-
-#endif // JS_CMP_GLOBAL_VOID_HPP
