+++ lib/global/IsNaN.cpp
@@ -1,4 +1,5 @@
-#include "../../includes/global/global_isNaN.hpp"
+#include "../../includes/global/globalFunctions.hpp"
+#include "../../includes/utils/Convert.hpp"
 
 bool global_isNaN(const JS::Any& number) {
     return std::isnan(JS::CONVERT::ToNumber(number));
