--- lib/global/global_isFinite.cpp
@@ -1,6 +1,0 @@
-#include "../../includes/global/global_isFinite.hpp"
-
-bool global_isFinite(const JS::Any& number) {
-    return std::isfinite(JS::CONVERT::ToNumber(number));
-}
-
