+++ lib/global/void.cpp
@@ -1,4 +1,5 @@
-#include "../../includes/global/global_void.hpp"
+#include "../../includes/global/globalFunctions.hpp"
+#include "../../includes/types/JsAny.hpp"
 
 JS::Undefined global_void(const JS::Any& args) {
     JS::Any value = args[0];
